+++ pyabc/sampler/redis_eps/cli.py
@@ -2,13 +2,13 @@ import logging
 import os
 import random
 import socket
-from multiprocessing import Process
 from time import time
 
 import click
 import numpy as np
 from redis import StrictRedis
 
+from ..util import get_mp_process
 from .cmd import (
     ANALYSIS_ID,
     DYNAMIC,
@@ -89,8 +89,9 @@ def work(
         return _work(host, port, runtime, password, catch)
 
     # define parallel processes
+    process_cls = get_mp_process()
     procs = [
-        Process(
+        process_cls(
             target=_work,
             args=(host, port, runtime, password, catch),
             daemon=daemon,
@@ -198,7 +199,8 @@ def _work(
             logger.info('Received stop signal. Shutdown redis worker.')
             return
 
-        # TODO other messages (some integers?) are ignored
+        # Any other messages on the channel (e.g. redis subscription-count
+        # notifications) are ignored.
 
         # check total time condition
         elapsed_time = time() - start_time