+++ pyabc/storage/bytes_storage.py
@@ -16,7 +16,7 @@ try:
                 py_object_ = conv.rpy2py(object_)
             # Ensure factor columns are converted to strings
             for col in py_object_.columns:
-                if isinstance(py_object_[col], pd.CategoricalDtype):
+                if isinstance(py_object_[col].dtype, pd.CategoricalDtype):
                     py_object_[col] = py_object_[col].astype(str)
             return py_object_
         return object_